Move company deactivation functionality and related pop-ups to component to be displayed in the navbar.
